requirements.txt
confluent-kafka==2.6.1
faker==25.9.1
pydantic==2.8.2
python-dotenv==1.0.1
redis==5.0.7
psycopg2-binary==2.9.9
fastapi==0.112.2
uvicorn==0.30.6
numpy>=2.0.2
pandas>=2.2.2
scikit-learn>=1.5.1
joblib>=1.4.2
